chore: Processed SonarCloud remarks
